tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/animation-delay_inherit.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - animation-delay: inherit</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#the-animation-delay-property-">
   <meta name="flags" content="dom">
   <meta name="assert" content="The computed value of 'animation-delay' is the
   same as the specified value">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes slide {
         from {
            margin-left: 0px;
         }
         to {
            margin-left: 100px;
         }
      }
      body > div {
         animation-delay: 1s;
      }
      div > div {
         width: 100px;
         height: 100px;
         background: blue;
         animation-duration: 0.5s;
         animation-fill-mode: forwards;
         animation-delay: inherit;
      }
      .test {
         animation-name: slide;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelectorAll('div')[1];
         var log = document.querySelectorAll('p')[1].firstChild;
         function aboutEqual(observed, actual) {
            var maxDeviation = 100;
            return (Math.abs(observed - actual) < maxDeviation);
         }
         var timeout = setTimeout(function() {
            test(function(){assert_true(false);}, testName); done();
            log.data = 'FAIL';
         }, 1500);
         div.addEventListener('animationstart', function() {
            clearTimeout(timeout);
            var delaySeen = Date.now() - delayStart;
            var innerStyle = getComputedStyle(div, null);
            var delaySpec = innerStyle.getPropertyValue('animation-delay');
            var passed = (delaySpec == '1s' && aboutEqual(delaySeen, 1000));
            test(function(){assert_true(passed);}, testName); done();
            log.data = (passed ? 'PASS if the animation occurred.' : 'FAIL');
         });
         var delayStart = Date.now();
         div.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The blue square below should start sliding right after one second.</p>
   <p>Testing...</p>
   <div>
      <div></div>
   </div>
</body>
</html>
